backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlmodel==0.0.22
pillow==11.1.0
rapidocr-onnxruntime==1.4.2
opencv-python-headless==4.11.0.86
numpy==2.2.2
